Refactor to new submodule
