+++ src/flowrep/compiler/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module handles the reverse transformation of flowrep objects back to python whence
+they can be parsed.
+"""
